+++ index.html
@@ -8,12 +8,12 @@
     <link rel="stylesheet" href="./index.css">
     <link rel="stylesheet" href="https://bootswatch.com/5/sketchy/bootstrap.min.css"/>
 </head>
-<script>
+<!-- <script>
     window.dataLayer = window.dataLayer || [];
     function gtag(){window.dataLayer.push(arguments);}
     gtag('js', new Date());
     gtag('config', 'G-HQREYRLMDN');
-</script>
+</script> -->
 <body> 
     <header>
         <nav class="navbar">
@@ -65,8 +65,8 @@
     </ul>
   </section>
   <footer class="footer navbar">
-    <p>&copy;Yordi & carensiya 2023 all right Reserved.</p>
+    <p>&copy; carensiya 2023 all right Reserved.</p>
   </footer>
-<script src="./assets/app.js"></script>
+<script src="/index.js" type="module"></script>
 </body>
 </html>
