+++ modules/books.js
@@ -1,0 +1,7 @@
+
+export default class Books {
+    constructor(title, author) {
+      this.title = title;
+      this.author = author;
+    }
+}
